Fix install with custom prefix in top level Makefile.
Pass the prefix given on the command line to Make invocations
for subdirectories.

Also, make 'install' target to depend on 'all' and also

This fixes issue#12 for now. Later we probably need to redo
the top level Makefile completely to make it behave better
with Makefiles in gnat_linux directory.