Fix static library build.
When Ahven is build as a static library, we need to add
linker option '-lgnarl' during the linking. Otherwise,
we get a lot of errors about missing functions when compiling
the actual test runner.

Also, define Library_Interface at the top level, so
the static build also has it. This fixes the include
dir population.